requirements.txt
google-genai>=0.2.0
streamlit>=1.28.0
python-dotenv>=1.0.0

